+++ server/auth/index.js
@@ -29,7 +29,7 @@ router.post(
 	'/login',
 	function(req, res, next) {
 		console.log(req.body)
-		console.log('================')
+		console.log("THIS IS THE BREAKING POINT *******");
 		next()
 	},
 	passport.authenticate('local'),
@@ -57,16 +57,16 @@ router.post('/logout', (req, res) => {
 
 router.post('/signup', (req, res) => {
 	console.log(req.body)	
-	const { username, password } = req.body
+	const { email, password } = req.body
 	// ADD VALIDATION
-	User.findOne({ 'local.username': username }, (err, userMatch) => {
+	User.findOne({ 'local.email': email }, (err, userMatch) => {
 		if (userMatch) {
 			return res.json({
-				error: `Sorry, already a user with the username: ${username}`
+				error: `Sorry, already a user with the email: ${email}`
 			})
 		}
 		const newUser = new User({
-			'local.username': username,
+			'local.email': email,
 			'local.password': password
 		})
 		newUser.save((err, savedUser) => {