+++ server/db/models/user.js
@@ -8,14 +8,13 @@ const userSchema = new Schema({
 	firstName: { type: String, unique: false },
 	lastName: { type: String, unique: false },
 	local: {
-		username: { type: String, unique: false, required: false },
+		email: { type: String, unique: false, required: false },
 		password: { type: String, unique: false, required: false }
 	},
 	google: {
 		googleId: { type: String, required: false }
 	},
 	phone: Number,
-	email: String,
 	workplace: { type : Schema.ObjectId, ref: 'Workplace' },
 	photos: []
 	// local: {