+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.0.8] - 2017-07-27
 ### Changed
-- changed from signing up, signing in with email to using a username
+- changed from signing up, signing in with email to using a email
 <!-- ## [Unreleased]
 
 ## [1.0.0] - 2017-06-20